reduced app deletage size
